fix: update build output path for staging build
