+++ .circleci/config.yml
@@ -114,7 +114,7 @@ jobs:
           command: |
             cd ./packages/self-service
             npm install
-            npm run build
+            npm run build version=${CIRCLE_TAG}
 #            TODO: update self-service to use yarn workspaces
 #            yarn install --immutable --immutable-cache
 #            yarn build
