+++ packages/self-service/package.json
@@ -5,8 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build:post": "node scripts/generate-index.js",
-    "build": "rm -rf dist && vite build && npm run build:post",
+    "build": "node scripts/build.js",
     "preview": "vite preview",
     "lint": "npx eslint src",
     "lint:fix": "npm run lint -- --fix",
