--- packages/self-service/scripts/widget-version.js
@@ -1,7 +1,0 @@
-import fs from 'node:fs';
-
-const packageJson = JSON.parse(fs.readFileSync('../../package.json', 'utf8'));
-
-export function getWidgetVersion() {
-  return packageJson.version;
-}
